--- turbo-repo/apps/app/src/features/dashboard/dashlets/common/index.ts
@@ -1,246 +1,0 @@
-/**
- * Common components for dashlet settings
- */
-
-export {
-  SettingsShell,
-  type SettingsTab,
-  type SettingsShellProps,
-  buildStandardTabs,
-} from "./settings-shell";
-export { useWidgetRefreshSettings } from "./use-widget-refresh-settings";
-export { SettingsTabButton } from "./settings-tab-button";
-export { useDataProvider } from "./use-data-provider";
-export {
-  SettingsTextField,
-  SettingsNumberField,
-  SettingsTextareaField,
-  SettingsSelectField,
-  SettingsFieldGrid,
-  SettingsPickerRow,
-  SettingsPickerItem,
-  SettingsTitleValueUnit,
-  HbTextField,
-  HbTextFieldList,
-  HbInlineInput,
-  HbTextareaField,
-} from "./settings-fields";
-export {
-  type HandlebarsStatus,
-  findHandlebarsExpressions,
-  getHandlebarsStatus,
-  getFlowbiteColor,
-  resolveDataProperty,
-} from "./handlebars-helpers";
-export {
-  type DataMode,
-  type ColumnType,
-  type TableColumn,
-  type SortConfig,
-} from "./column-types";
-export {
-  getBadgeClasses,
-  getProgressColor,
-  renderProgress,
-  getSignedClasses,
-  renderCell,
-} from "./cell-renderers";
-export { type PillProps, Pill } from "./pill";
-export {
-  type ColumnItem,
-  toColumnItems,
-  fromColumnItems,
-} from "./column-helpers";
-export { useDynamicRows } from "./use-dynamic-rows";
-export { usePgrestRows } from "./use-pgrest-rows";
-export {
-  type PgrestParam,
-  type PgrestHttpMethod,
-  type PgrestParamItem,
-  humanizeKey,
-  toPgrestParamItems,
-  fromPgrestParamItems,
-  EMPTY_PGREST_PARAMS,
-} from "./pgrest-types";
-export { parseRows, buildPgrestFetch } from "./pgrest-utils";
-export {
-  type PgrestSettingsStateConfig,
-  usePgrestSettingsState,
-} from "./use-pgrest-settings-state";
-export { PgrestSettingsSection } from "./pgrest-settings-section";
-export { PgrestFunctionAutocomplete } from "./pgrest-function-autocomplete";
-export { type FilterItemConfig, type FilterConfig } from "./filter-types";
-export {
-  type FilterItem,
-  toFilterItems,
-  fromFilterItems,
-  normalizeFilterConfig,
-} from "./filter-helpers";
-export { FilterPillRow } from "./filter-pill-row";
-export {
-  type UseFilterAndSortResult,
-  useFilterAndSort,
-} from "./use-filter-and-sort";
-export {
-  ColumnEditor,
-  FilterEditor,
-  SortEditor,
-  DataProviderTab,
-  CheckboxColumnList,
-} from "./settings-sections";
-export {
-  type SettingsStateConfig,
-  useSettingsState,
-} from "./use-settings-state";
-export { TableListSettingsShell } from "./table-list-settings-shell";
-export {
-  type TemplateField,
-  compileTemplates,
-  resolveTemplate,
-  resolveHandlebarsField,
-  buildDataProviderContext,
-} from "./use-handlebars-templates";
-export { registerHandlebarsHelpers } from "./register-handlebars-helpers";
-export { SuggestionInput } from "./suggestion-input";
-export { useCompiledColumns } from "./use-compiled-columns";
-export {
-  buildPgrestSettingsConfig,
-  buildPgrestContentLabels,
-  defaultOnColumnsDetected,
-  buildSimplePgrestConfig,
-} from "./pgrest-settings-helpers";
-export { IconColorPickerRow } from "./icon-color-picker-row";
-export { DASHLET_ICON_OPTIONS, type DashletIconKey } from "./icon-options";
-export { usePgrestResolvedFields } from "./use-pgrest-resolved-fields";
-export { PgrestDataTab } from "./pgrest-data-tab";
-export { usePlannerData } from "./use-planner-data";
-export { useDashletData } from "./use-dashlet-data";
-export { PlannerVariableSelector } from "./planner-variable-selector";
-export { useActiveProviders } from "./use-active-providers";
-export {
-  DashletLoading,
-  DashletError,
-  parseResolvedNumber,
-} from "./dashlet-states";
-export { DataProviderEntries } from "./data-provider-entries";
-export {
-  type SimpleDataMode,
-  isRemoteDataMode,
-  useSimplePgrestSettings,
-} from "./use-simple-pgrest-settings";
-export {
-  type PgrestDashletFields,
-  useDashletPgrest,
-  useHybridPgrestContext,
-} from "./use-dashlet-pgrest";
-export {
-  type SettingsFieldDef,
-  type SimpleDashletSettingsProps,
-  SimpleDashletSettings,
-  createSimpleDashletSettings,
-  createSettingsField,
-  useFieldState,
-} from "./simple-dashlet-settings";
-export {
-  type ThresholdTarget,
-  type ThresholdRule,
-  type ThresholdConfig,
-  type ThresholdRuleItem,
-  THRESHOLD_TARGETS,
-  THRESHOLD_TARGET_LABELS,
-} from "./threshold-types";
-export {
-  evaluateThreshold,
-  getThresholdBgClasses,
-  getThresholdTextClasses,
-  getThresholdTextStyle,
-  getThresholdIconClasses,
-  getThresholdBorderClasses,
-  getThresholdStrokeClass,
-  getThresholdStrokeStyle,
-  getThresholdBarClass,
-  getThresholdBarStyle,
-  getThresholdGradientClasses,
-  isLegacyColor,
-  isHexColor,
-} from "./threshold-engine";
-export {
-  DEFAULT_THRESHOLD_CONFIG,
-  normalizeThresholdConfig,
-  toThresholdRuleItems,
-  fromThresholdRuleItems,
-} from "./threshold-helpers";
-export {
-  type ThresholdResult,
-  type ThresholdOptions,
-  useThreshold,
-  useRowThreshold,
-} from "./use-threshold";
-export { ThresholdEditor } from "./threshold-editor";
-export { useThresholdSettings } from "./use-threshold-settings";
-export { DeleteItemButton } from "./delete-item-button";
-export { buildCsvContent, downloadCsv } from "./export-csv";
-export { ExportDropdown } from "./export-dropdown";
-export { usePortalDropdown } from "./use-portal-dropdown";
-export { DashletTitleBar } from "./dashlet-title-bar";
-export {
-  type CompareMode,
-  type TargetOption,
-  type CompareFieldOption,
-  type ColorRule,
-  type ColorRuleItem,
-  type ColorRulesConfig,
-  type NormalizeOptions,
-  type UseColorRuleSettingsOptions,
-  type ColorRuleSetterProps,
-  normalizeColorRulesConfig,
-  useColorRuleSettings,
-  ColorRuleSetter,
-} from "./color-rule-setter";
-export {
-  type ColumnOption,
-  type RowColorRuleItem,
-  type RowColorRuleSetterProps,
-  RowColorRuleSetter,
-} from "./row-color-rule-setter";
-export {
-  type TargetConfig,
-  type ValueColorRulesFactoryConfig,
-  createValueColorRulesUtils,
-  textIconTargets,
-  textBgTargets,
-  textBgIconTargets,
-  borderIconTextTargets,
-  textBarBadgeTargets,
-} from "./value-color-rules-factory";
-export {
-  isGreaterOperator,
-  isLessOperator,
-  type SortableRule,
-  type ColorableRule,
-  type EvaluatableRule,
-  type ComparableRule,
-  type EvaluatedColors,
-  sortColorRules,
-  sortColorRulesWithFields,
-  getCompareValue,
-  evaluateColorRulesGeneric,
-  evaluateColorRulesWithFields,
-  hexToRgba,
-  buildTextStyle,
-  buildBgStyle,
-  buildIconStyle,
-  getConditionalClasses,
-  getBadgeClasses as getStatBadgeClasses,
-} from "./color-rule-evaluation";
-export {
-  ColorRuleRow,
-  OperatorDropdown,
-  ValueInput,
-  DeleteButton,
-  ToggleSectionHeader,
-  AddRuleButton,
-  ColumnDropdown,
-  type ColumnDropdownOption,
-  DEFAULT_COLOR_PRESETS,
-} from "./color-rule-row";
