+++ turbo-repo/apps/app/src/features/dashboard/dashlets/stat_expandable/dashlet.tsx
@@ -3,15 +3,14 @@
 import { useState } from "react";
 import { HiChevronDown, HiChevronUp } from "react-icons/hi2";
 import type { DashletComponentProps, DashletLayoutDefaults } from "../types";
-import type { PgrestDashletFields } from "../common";
+import { type PgrestDashletFields } from "../common/use-dashlet-pgrest";
+import { useDashletPgrest } from "../common/use-dashlet-pgrest";
+import { DashletLoading, DashletError } from "../common/dashlet-states";
 import {
-  useDashletPgrest,
-  DashletLoading,
-  DashletError,
   evaluateColorRulesGeneric,
   buildTextStyle,
   getConditionalClasses,
-} from "../common";
+} from "../common/color-rule-evaluation";
 import { useEffectiveRefreshInterval } from "../../hooks/use-effective-refresh-interval";
 import type {
   ValueColorRulesConfig,
