+++ turbo-repo/apps/app/src/features/dashboard/dashlets/stat_progress/value-color-rules.tsx
@@ -6,10 +6,12 @@ import {
   type ColorRule,
   type ColorRuleItem,
   type ColorRulesConfig,
+} from "../common/color-rule-setter";
+import {
   normalizeColorRulesConfig,
   ColorRuleSetter,
   useColorRuleSettings,
-} from "../common";
+} from "../common/color-rule-setter";
 
 // ============================================================================
 // Types specific to stat_progress
