+++ turbo-repo/apps/app/src/features/dashboard/dashlets/stat_sensitive/dashlet.tsx
@@ -3,8 +3,9 @@
 import { useState } from "react";
 import { HiEye, HiEyeSlash } from "react-icons/hi2";
 import type { DashletComponentProps, DashletLayoutDefaults } from "../types";
-import type { PgrestDashletFields } from "../common";
-import { useDashletPgrest, DashletLoading, DashletError } from "../common";
+import { type PgrestDashletFields } from "../common/use-dashlet-pgrest";
+import { useDashletPgrest } from "../common/use-dashlet-pgrest";
+import { DashletLoading, DashletError } from "../common/dashlet-states";
 import { useEffectiveRefreshInterval } from "../../hooks/use-effective-refresh-interval";
 import { useRowThreshold } from "../common/use-threshold";
 import { getThresholdTextClasses } from "../common/threshold-engine";
