+++ turbo-repo/apps/app/src/features/dashboard/dashlets/stat_sparkline/dashlet.settings.tsx
@@ -3,7 +3,8 @@
 import { useState } from "react";
 import type { DashletSettingsProps } from "../types";
 import type { DashletConfig } from "./dashlet";
-import { SimpleDashletSettings, SettingsTextField } from "../common";
+import { SimpleDashletSettings } from "../common/simple-dashlet-settings";
+import { SettingsTextField } from "../common/settings-fields";
 import { tr } from "@/features/i18n/tr.service";
 
 const FIELDS = [
