+++ turbo-repo/apps/app/src/features/dashboard/dashlets/stat_status/dashlet.tsx
@@ -6,14 +6,10 @@ import type {
   DashletLayoutDefaults,
   DataProviderEntry,
 } from "../types";
-import type { PgrestDashletFields } from "../common";
-import {
-  useHybridPgrestContext,
-  DashletLoading,
-  DashletError,
-  evaluateColorRulesGeneric,
-  getConditionalClasses,
-} from "../common";
+import { type PgrestDashletFields } from "../common/use-dashlet-pgrest";
+import { useHybridPgrestContext } from "../common/use-dashlet-pgrest";
+import { DashletLoading, DashletError } from "../common/dashlet-states";
+import { evaluateColorRulesGeneric, getConditionalClasses } from "../common/color-rule-evaluation";
 import { useEffectiveRefreshInterval } from "../../hooks/use-effective-refresh-interval";
 import { resolveHandlebarsField } from "../common/use-handlebars-templates";
 import type {
