+++ turbo-repo/apps/web-site/.flowbite-react/class-list.json
@@ -1,13 +1,10 @@
 [
-  "-translate-x-1/2",
-  "-translate-y-1/2",
   "2xl",
   "3xl",
   "4xl",
   "5xl",
   "6xl",
   "7xl",
-  "absolute",
   "bg-blue-700",
   "bg-cyan-700",
   "bg-gray-700",
@@ -23,10 +20,7 @@
   "bg-teal-700",
   "bg-transparent",
   "bg-white",
-  "bg-white/30",
-  "bg-white/50",
   "bg-yellow-400",
-  "block",
   "border",
   "border-b",
   "border-b-0",
@@ -47,21 +41,16 @@
   "border-t",
   "border-teal-700",
   "border-yellow-400",
-  "bottom-5",
   "bottom-center",
   "bottom-left",
   "bottom-right",
   "center-left",
   "center-right",
-  "cursor-default",
-  "cursor-grab",
   "dark:bg-blue-600",
   "dark:bg-cyan-600",
   "dark:bg-gray-600",
   "dark:bg-gray-700",
   "dark:bg-gray-800",
-  "dark:bg-gray-800/30",
-  "dark:bg-gray-800/50",
   "dark:bg-gray-900/80",
   "dark:bg-green-600",
   "dark:bg-indigo-600",
@@ -99,14 +88,11 @@
   "dark:focus:ring-red-800",
   "dark:focus:ring-teal-800",
   "dark:focus:ring-yellow-900",
-  "dark:group-focus:ring-gray-800/70",
-  "dark:group-hover:bg-gray-800/60",
   "dark:hidden",
   "dark:hover:bg-blue-700",
   "dark:hover:bg-cyan-700",
   "dark:hover:bg-gray-600",
   "dark:hover:bg-gray-700",
-  "dark:hover:bg-gray-800",
   "dark:hover:bg-green-700",
   "dark:hover:bg-indigo-700",
   "dark:hover:bg-lime-700",
@@ -133,7 +119,6 @@
   "dark:text-blue-500",
   "dark:text-cyan-500",
   "dark:text-gray-400",
-  "dark:text-gray-800",
   "dark:text-green-500",
   "dark:text-indigo-400",
   "dark:text-lime-500",
@@ -169,13 +154,8 @@
   "focus:ring-yellow-300",
   "font-medium",
   "gap-4",
-  "group-focus:outline-none",
-  "group-focus:ring-4",
-  "group-focus:ring-white",
-  "group-hover:bg-white/50",
   "h-10",
   "h-12",
-  "h-3",
   "h-5",
   "h-8",
   "h-9",
@@ -198,7 +178,6 @@
   "hover:bg-purple-800",
   "hover:bg-red-800",
   "hover:bg-teal-800",
-  "hover:bg-white",
   "hover:bg-yellow-500",
   "hover:border-blue-800",
   "hover:border-cyan-800",
@@ -226,8 +205,6 @@
   "justify-end",
   "justify-start",
   "last:rounded-e-lg",
-  "left-0",
-  "left-1/2",
   "max-h-[90dvh]",
   "max-w-2xl",
   "max-w-3xl",
@@ -252,9 +229,7 @@
   "opacity-50",
   "overflow-auto",
   "overflow-x-hidden",
-  "overflow-x-scroll",
   "overflow-y-auto",
-  "overflow-y-hidden",
   "p-1.5",
   "p-2",
   "p-2.5",
@@ -264,17 +239,50 @@
   "pointer-events-none",
   "pt-0",
   "px-3",
-  "px-4",
   "px-5",
   "px-6",
   "relative",
-  "right-0",
   "rounded-b",
   "rounded-full",
   "rounded-lg",
   "rounded-md",
   "rounded-none",
   "rounded-t",
+  "rounded-t-lg",
+  "shadow-md",
+  "shadow-sm",
+  "shadow-xs",
+  "space-x-2",
+  "text-base",
+  "text-blue-700",
+  "text-center",
+  "text-cyan-700",
+  "text-gray-400",
+  "text-gray-500",
+  "text-gray-700",
+  "text-gray-800",
+  "text-gray-900",
+  "text-green-700",
+  "text-indigo-700",
+  "text-lime-700",
+  "text-pink-700",
+  "text-primary-700",
+  "text-purple-700",
+  "text-red-700",
+  "text-sm",
+  "text-teal-700",
+  "text-white",
+  "text-xl",
+  "text-xs",
+  "text-yellow-400",
+  "top-0",
+  "top-center",
+  "top-left",
+  "top-right",
+  "w-5",
+  "w-full",
+  "z-50"
+],
   "rounded-t-lg",
   "scroll-smooth",
   "shadow-md",
